requirements.txt
stable-diffusion-cpp-python
huggingface_hub
runpod
Pillow